+++ client/src/components/StudentMentorship.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { Message, Form, Modal, Menu, Label, Card, Grid, Image, Button } from 'semantic-ui-react';
+import { Modal, Form, Menu, Label, Card, Grid, Image, Button, Message } from 'semantic-ui-react';
 import { makeCall } from '../apis'
 import swal from 'sweetalert'
 
@@ -32,57 +32,45 @@ export const timeSlotOptions = [
 
 /*
  * DETAILS:
- * Parent component, contains menu and switches out active RequestCards
+ * Parent component, contains menu and switches out active SchedulingCards
  * PROPS
  * userDetails - full profile of logged in user
  */
-export default class RequestsView extends Component {
+export default class StudentMentorship extends Component {
     constructor(props){
         super(props)
         this.state={
-            activeItem: 'unconfirmed',
+            activeItem: 'confirmed',
             unconfirmed: [],
             confirmed: [],
             completed: [],
             timeOffset: 0,
-            confirmedTimes: [],
             userDetails: null
         }
         this.handleStatusUpdate = this.handleStatusUpdate.bind(this)
     }
     
-    async handleStatusUpdate(requests) {
-        if (requests.requests === []) return;
-        let confirmedTimes = await this.populateConfirmedTimes(requests.requests[1])
+    async handleStatusUpdate(schedulings) {
+        if (schedulings.schedulings === []) return;
         this.setState({
-            unconfirmed: requests.requests[0],
-            confirmed: requests.requests[1],
-            completed: requests.requests[2],
-            confirmedTimes: confirmedTimes,
+            unconfirmed: schedulings.schedulings[0],
+            confirmed: schedulings.schedulings[1],
+            completed: schedulings.schedulings[2],
         })
     }
 
     async componentWillMount() {
         let timeOffset = this.props.userDetails.timeZone
-        let requests = await this.getRequests(timeOffset)
-        let confirmedTimes = await this.populateConfirmedTimes(requests.requests[1])
+        let schedulings = await this.getSchedulings(timeOffset)
         this.setState({
             timeOffset: timeOffset,
-            unconfirmed: requests.requests[0],
-            confirmed: requests.requests[1],
-            completed: requests.requests[2],
-            confirmedTimes: confirmedTimes
+            unconfirmed: schedulings.schedulings[0],
+            confirmed: schedulings.schedulings[1],
+            completed: schedulings.schedulings[2],
         })
     }
-
-    getRequests(timeOffset) {
-        return makeCall({}, '/request/getRequests/'+this.props.userDetails._id+'/'+timeOffset, 'get')
-    }
-
-    populateConfirmedTimes(requests) {
-       return requests.map(confirmedRequest => {
-            return (confirmedRequest.time[0].id)
-        })
+    getSchedulings(timeOffset) {
+        return makeCall({}, '/request/getSchedulings/'+this.props.userDetails._id +'/'+ this.props.role +'/' + timeOffset, 'get')
     }
 
     handleMenuClick = (e, { id }) => this.setState({ activeItem: id })
@@ -91,17 +79,6 @@ export default class RequestsView extends Component {
         return(
             <div>
             <Menu secondary stackable>
-                <Menu.Item
-                    id='unconfirmed'
-                    name='Unconfirmed Meetings'
-                    active={this.state.activeItem === 'unconfirmed'}
-                    onClick={this.handleMenuClick}
-                >
-                    Unconfirmed Meetings
-                    {   (this.state.unconfirmed !== []) &&
-                        <Label color='teal'>{this.state.unconfirmed.length}</Label>
-                    }
-                </Menu.Item>
                 <Menu.Item
                     id='confirmed'
                     name='Confirmed Meetings'
@@ -114,43 +91,56 @@ export default class RequestsView extends Component {
                     }
                 </Menu.Item>
                 <Menu.Item
+                    id='unconfirmed'
+                    name='Unconfirmed Meetings'
+                    active={this.state.activeItem === 'unconfirmed'}
+                    onClick={this.handleMenuClick}
+                >
+                    Unconfirmed Meetings
+                    {   (this.state.unconfirmed !== []) &&
+                        <Label color='teal'>{this.state.unconfirmed.length}</Label>
+                    }
+                </Menu.Item>
+                <Menu.Item
                     id='completed'
                     name='Completed Meetings'
                     active={this.state.activeItem === 'completed'}
                     onClick={this.handleMenuClick}
                 />
             </Menu>
-            {this.state.activeItem === 'unconfirmed' &&
+            {this.state.activeItem === 'confirmed' &&
                 <div style={{paddingLeft: 13, paddingRight: 13}}>
-                    <RequestCards 
-                        confirmedTimes={this.state.confirmedTimes}
+                    <SchedulingCards 
                         activeSet={this.state.activeItem}
-                        requests={this.state.unconfirmed}
-                        liftRequests={this.handleStatusUpdate}
+                        schedulings={this.state.confirmed}
+                        liftSchedulings={this.handleStatusUpdate}
                         timeOffset={this.state.timeOffset}
                         userId={this.props.userDetails._id}
+                        userRole={this.props.role}
                     />
                 </div>
             }
-            {this.state.activeItem === 'confirmed' &&
+            {this.state.activeItem === 'unconfirmed' &&
                 <div style={{paddingLeft: 13, paddingRight: 13}}>
-                    <RequestCards 
+                    <SchedulingCards 
                         activeSet={this.state.activeItem}
-                        requests={this.state.confirmed}
-                        liftRequests={this.handleStatusUpdate}
+                        schedulings={this.state.unconfirmed}
+                        liftSchedulings={this.handleStatusUpdate}
                         timeOffset={this.state.timeOffset}
                         userId={this.props.userDetails._id}
+                        userRole={this.props.role}
                     />
                 </div>
             }
             {this.state.activeItem === 'completed' &&
                 <div style={{paddingLeft: 13, paddingRight: 13}}>
-                    <RequestCards 
-                        liftRequests={this.handleStatusUpdate}
-                        userId={this.props.userDetails._id}
-                        timeOffset={this.state.timeOffset}
+                    <SchedulingCards 
                         activeSet={this.state.activeItem}
-                        requests={this.state.completed}
+                        schedulings={this.state.completed}
+                        liftSchedulings={this.handleStatusUpdate}
+                        timeOffset={this.state.timeOffset}
+                        userId={this.props.userDetails._id}
+                        userRole={this.props.role}
                     />
                 </div>
             }
@@ -162,71 +152,71 @@ export default class RequestsView extends Component {
 
 /*
  * DETAILS
- * Child component - takes array of requests, constructs display
+ * Child component - takes array of schedulings, constructs display
  *                   shows valid actions for a given set
  * PROPS
  * Always:
  * activeSet (string) - Shows what set is currently being used
- * requests - array of relevant requests
+ * schedulings - array of relevant requests
  * 
  * Sometimes:
  * userId - mongo ID for profile (used to update records)
- * liftRequests (method) - lifts results of an update to parent component
+ * userRole - user's role in the mongoDB
+ * liftSchedulings (method) - lifts results of an update to parent component
  * timeOffset
- * confirmedTimes - array of meeting times that have already been confirmed
- *                  used to disable approval, preventing double booking
  */
-class RequestCards extends Component {
+class SchedulingCards extends Component {
     state={
-        requests: [],
+        schedulings: [],
         display: [],
-        showFeedbackModal: false,
-        requestDetails: null,
-        finalNote: ''
+        schedulingDetails: [],
+        feedback: '',
+        showFeedbackModal: false
     }
     // This allows the component to update its state should a prop value change
-    async componentWillReceiveProps({requests}) {
-        await this.setState({requests: requests})
-        this.constructDisplay(this.state.requests)
+    async componentWillReceiveProps({schedulings}) {
+        await this.setState({schedulings: schedulings})
+        this.constructDisplay(this.state.schedulings)
     }
     // This ensures that the component doesn't use an old prop on menu change
     componentWillMount() {
-        this.constructDisplay(this.props.requests)
+        this.constructDisplay(this.props.schedulings)
     }
 
-    constructRequest(request) {
+    constructRequest(scheduling) {
+        const cardHeader = (this.props.activeSet !== 'completed'? 'Meeting with: ' : 'Completed call with: ')
         return (
-            <Grid key={request._id} columns={'equal'}>
+            <Grid key={scheduling._id} columns={'equal'}>
             <Grid.Row columns={2}>
                 <Grid.Column width={4}>
                     <Image
                         size='small'
                         centered
                         rounded
-                        src={request.requesterObj.imageURL}
+                        src={scheduling.mentor.imageURL}
                     />
                 </Grid.Column>
                 <Grid.Column>
                     <Card fluid>
                         <Card.Content>
                             <Card.Header>
-                                Request from {request.requesterObj.name} ({request.requesterRole.toLowerCase()})
+                                {cardHeader} {scheduling.mentor.name}
                             </Card.Header>           
-                            <Card.Meta>{request.status}</Card.Meta>
-                            <Card.Description>Topic: {request.topic}</Card.Description>
-                            <Card.Description>Time: {request.time[0].day} from {timeSlotOptions[request.time[0].time/100]}</Card.Description>
-                            { request.note &&
-                                <Card.Description>Note from requester: {request.note}</Card.Description>
+                            <Card.Meta>{scheduling.status}</Card.Meta>
+                            <Card.Description>Topic: {scheduling.topic}</Card.Description>
+                            <Card.Description>Time: {scheduling.time[0].day} from {timeSlotOptions[scheduling.time[0].time/100]}</Card.Description>
+                            {scheduling.studentNote &&
+                                <Card.Description>Your Note: {scheduling.studentNote}</Card.Description>
                             }
-                            { request.finalNote &&
-                                <Card.Description>Final note from mentor: {request.finalNote}</Card.Description>
+                            {scheduling.finalNote &&
+                                <Card.Description>Final note from mentor: {scheduling.finalNote}</Card.Description>
                             }
-                            { request.feedback &&
-                                <Card.Description>Feedback for mentor: {request.feedback}</Card.Description>
+                            {scheduling.feedback &&
+                                <Card.Description>Feedback for mentor: {scheduling.feedback}</Card.Description>
                             }
                             <br />
                         </Card.Content>
-                        {this.buttonDisplay(request)}
+                            {this.buttonDisplay(scheduling)}
                     </Card>
                 </Grid.Column>
             </Grid.Row>
@@ -234,71 +224,62 @@ class RequestCards extends Component {
         )
     }
 
-    buttonDisplay(request) {
+    buttonDisplay(scheduling) {
         if (this.props.activeSet === 'unconfirmed') {
-            let disableApprove = this.props.confirmedTimes.includes(request.time[0].id)
             return (
-                <Button.Group>
-                    <Button
-                        positive
-                        requestid={request._id}
-                        newstatus={'Confirmed'}
-                        onClick={this.handleClick.bind(this)}
-                        disabled={disableApprove}
-                    >
-                        Approve
-                    </Button>
-                    <Button
-                        negative
-                        requestid={request._id}
-                        newstatus={'Rejected'}
-                        onClick={this.handleClick.bind(this)}
-                    >
-                        Reject
-                    </Button>
-                </Button.Group>
+                <Button
+                    fluid
+                    negative
+                    requestid={scheduling._id}
+                    newstatus={'Rejected'}
+                    onClick={this.handleClick.bind(this)}
+                >
+                    Cancel
+                </Button>
             )
         } else if (this.props.activeSet === 'confirmed') {
             return(
+                <>
                 <Button.Group compact>
                     <Button 
                         color='blue' 
                         as='a'
-                        href={request.zoomLink} 
+                        href={scheduling.zoomLink} 
                         target="_blank"
                         rel="noopener noreferrer"
                     >
                         Join Call
                     </Button>
-                    <Button 
-                        positive
-                        requestid={request._id}
-                        newstatus={'Completed'}
+                    <Button
+                        negative
+                        requestid={scheduling._id}
+                        newstatus={'Rejected'}
                         onClick={this.handleClick.bind(this)}
                     >
-                        Mark Completed!
+                        Cancel
                     </Button>
                 </Button.Group>
+                </>
             )
         } else if (this.props.activeSet === 'completed') {
             return(
                 <Button
                     color='teal'
-                    requestId={request._id}
+                    requestid={scheduling._id}
                     onClick={this.toggleFeedbackModal.bind(this)}
                 >
-                    Leave a note!
+                    Provide feedback!
                 </Button>
             )
         }
     }
 
     async handleClick(e) {
-        let requests = await makeCall({
+        let schedulings = await makeCall({
                 requestId: e.currentTarget.getAttribute('requestid'),
                 newStatus: e.currentTarget.getAttribute('newstatus')
-            }, '/request/updateRequest/' + this.props.userId + '/' + this.props.timeOffset, 'patch');
-        if (!requests || requests.error) {
+            }, '/request/updateScheduling/' + this.props.userId + '/' + this.props.userRole + '/' + this.props.timeOffset, 'patch');
+        if (!schedulings || schedulings.error) {
             swal({
                 title: "Error!",
                 text: "There was an error updating this request, please try again.",
@@ -310,31 +291,31 @@ class RequestCards extends Component {
                 text: "Successfully updated this request!",
                 icon: "success"
             }).then(() => {
-                this.props.liftRequests(requests)
+                this.props.liftSchedulings(schedulings)
             })
         }
     }
 
     toggleFeedbackModal(e) {
-        let requestDetails = this.props.requests.find(request => request._id === e.currentTarget.getAttribute('requestid'))
-        let finalNote = null
-        if (requestDetails) {
-            finalNote = requestDetails.finalNote
+        let schedulingDetails = this.props.schedulings.find(request => request._id === e.currentTarget.getAttribute('requestid'))
+        let feedback = null
+        if (schedulingDetails) {
+            feedback = schedulingDetails.feedback
         }
         this.setState({
             showFeedbackModal: !this.state.showFeedbackModal,
-            requestDetails: requestDetails,
-            finalNote: finalNote
+            schedulingDetails: schedulingDetails,
+            feedback: feedback
         })
     }
 
     async submitFinalNote(e) {
         let requests = await makeCall({
-            requestId: this.state.requestDetails._id,
-            finalNote: this.state.finalNote
-        }, `/request/leaveFinalNote/${this.props.userId}/${this.props.timeOffset}`, 'patch')
+            requestId: this.state.schedulingDetails._id,
+            feedback: this.state.feedback
+        }, `/request/leaveFeedback/${this.props.userId}/${this.props.userRole}/${this.props.timeOffset}`, 'patch')
         this.setState({showFeedbackModal: !this.state.showFeedbackModal})
-        this.props.liftRequests(requests)
+        this.props.liftSchedulings(requests)
     }
 
     handleValueChange(e, {name, value}) {
@@ -344,10 +325,10 @@ class RequestCards extends Component {
         })
     }
 
-    constructDisplay(requests) {
+    constructDisplay(schedulings) {
         let display = []
-        for (let request of requests) {
-            display.push(this.constructRequest(request))
+        for (let scheduling of schedulings) {
+            display.push(this.constructRequest(scheduling))
         }
         this.setState({display: display})
     }
@@ -357,7 +338,7 @@ class RequestCards extends Component {
             <>
             {this.state.showFeedbackModal && 
                 <Modal open={this.state.showFeedbackModal}>
-                    <Modal.Header>Leave a final note for {this.state.requestDetails.requesterObj.name}</Modal.Header>
+                    <Modal.Header>Provide feedback for {this.state.schedulingDetails.mentor.name}</Modal.Header>
                     <Modal.Content>
                     <Grid stackable>
                         <Grid.Row columns={"equal"}>
@@ -365,18 +346,18 @@ class RequestCards extends Component {
                                 <Image
                                     floated='left'
                                     size='small'
-                                    src={this.state.requestDetails.requesterObj.imageURL}
+                                    src={this.state.schedulingDetails.mentor.imageURL}
                                     rounded
                                 />
                             </Grid.Column>
                             <Grid.Column>
                                 <Form>
                                     <Form.TextArea 
-                                        label={'Leave a note for ' + this.state.requestDetails.requesterObj.name + ':'}
-                                        placeholder='Provide a recap or leave any other useful notes here!'
+                                        label={'Leave a note for ' + this.state.schedulingDetails.mentor.name + ':'}
+                                        placeholder="How was your mentor able to help you? Is there something the mentor could do to be more helpful?"
                                         onChange={this.handleValueChange.bind(this)}
-                                        value={this.state.finalNote}
-                                        name='finalNote'
+                                        value={this.state.feedback}
+                                        name='feedback'
                                     />
                                 </Form>
                             </Grid.Column>
@@ -398,11 +379,11 @@ class RequestCards extends Component {
                 <Message.Header>No {this.props.activeSet} meetings!</Message.Header>
                 {
                     this.props.activeSet === 'unconfirmed' &&
-                    <Message.Content>Check back later!</Message.Content>
+                    <Message.Content>Make new schedulings in the alumni directory!</Message.Content>
                 }
                 {
                     this.props.activeSet === 'confirmed' &&
-                    <Message.Content>Confirm requests in the unconfirmed meetings tab!</Message.Content>
+                    <Message.Content>Check back later!</Message.Content>
                 }
                 </Message>
             }
